adds sweetalert to delete function

+++ server/public/scripts/client.js
@@ -1,5 +1,6 @@
 console.log('client sourced')
 
+
 $(document).ready(pageReady)
 
 function pageReady() {
@@ -18,14 +19,36 @@ function deleteTask() {
     console.log('delete clicked')
     const id = ($(this).data('id'))
 
-    $.ajax({
-        type: 'DELETE',
-        url: `/tasks/${id}`,
-    }).then(function (response) {
-        getTasks()
-    }).catch(function (error) {
-        console.log(error)
-    })
+    swal({
+        title: "Are you sure?",
+        text: "Once deleted, you will not be able to recover this imaginary file!",
+        icon: "warning",
+        buttons: true,
+        dangerMode: true,
+    }).then((willDelete) => {
+        if (willDelete) {
+            swal("Your task has been deleted", {
+                icon: "success",
+            });
+            $.ajax({
+                type: 'DELETE',
+                url: `/tasks/${id}`,
+            }).then(function (response) {
+                getTasks()
+            }).catch(function (error) {
+                console.log(error)
+            })
+        }
+    });
+
+    //     $.ajax({
+    //         type: 'DELETE',
+    //         url: `/tasks/${id}`,
+    //     }).then(function (response) {
+    //         getTasks()
+    //     }).catch(function (error) {
+    //         console.log(error)
+    //     })
 }
 
 function getTasks() {
@@ -54,7 +77,7 @@ function renderTasks(tasksToRender) {
         tr.append(`<td><input type="checkbox" id=${task.id} class="checkbox" data-id=${task.id}></td>
         <td class="task">${task.task}</td>
         <td><button class="deleteTaskBtn btn btn-outline-secondary" data-id=${task.id}>X</button></td>`)
-        
+
         $('#taskList').append(tr)
         $(`#${id}`).prop('checked', task.completed)
 
@@ -65,59 +88,59 @@ function renderTasks(tasksToRender) {
     }
 }
 
-    function addTask() {
-        console.log('add task clicked')
+function addTask() {
+    console.log('add task clicked')
 
-        let newTask = {
-            task: $('#taskIn').val(),
-            dueDate: $('#dueDateIn').val()
-        }
-        
-        console.log(newTask)
-
-        $('#taskIn').val('')
-        $('#dueDateIn').val('')
-
-        $.ajax({
-            type: 'POST',
-            url: '/tasks',
-            data: newTask
-        }).then(function (response) {
-            console.log(response)
-            getTasks(response)
-        }).catch(function (error) {
-            console.log(error)
-        })
-
+    let newTask = {
+        task: $('#taskIn').val(),
+        dueDate: $('#dueDateIn').val()
     }
 
-    function toggleComplete() {
-        console.log('checkbox checked')
-        console.log($(this).data('id'))
-        const id = $(this).data('id')
-        let newStatus = {}
+    console.log(newTask)
 
-        console.log($(this).is(':checked'))
-        if($(this).is(':checked')) {
-            newStatus = {
-                completed: true
-            }
-        } else {
-            newStatus = {
-                completed: false
-            }
+    $('#taskIn').val('')
+    $('#dueDateIn').val('')
+
+    $.ajax({
+        type: 'POST',
+        url: '/tasks',
+        data: newTask
+    }).then(function (response) {
+        console.log(response)
+        getTasks(response)
+    }).catch(function (error) {
+        console.log(error)
+    })
+
+}
+
+function toggleComplete() {
+    console.log('checkbox checked')
+    console.log($(this).data('id'))
+    const id = $(this).data('id')
+    let newStatus = {}
+
+    console.log($(this).is(':checked'))
+    if ($(this).is(':checked')) {
+        newStatus = {
+            completed: true
         }
-
-        console.log(newStatus)
-
-        $.ajax({
-            type: 'PUT',
-            url: `/tasks/${id}`,
-            data: newStatus
-        }).then(function (response) {
-            getTasks(response)
-        }).catch(function (error) {
-            console.log(error)
-        })
+    } else {
+        newStatus = {
+            completed: false
+        }
     }
+
+    console.log(newStatus)
+
+    $.ajax({
+        type: 'PUT',
+        url: `/tasks/${id}`,
+        data: newStatus
+    }).then(function (response) {
+        getTasks(response)
+    }).catch(function (error) {
+        console.log(error)
+    })
+}
